Updating original Appendix 1
A reviewer wanted to see stats about the representativeness of the
samples, so I’m re-including what was originally Appendix 1.  However,
since the original submission, the data structure has changed a bit (a
LOT more streamline) and multiple doesn’t work, so the graphs need to
be updated.

This commit only has the initial changes to make the basic plots for
ACRU.  Still need to make it work so it can be a facet_wrap statement